Fix: ConfigFile not found
The error only happens when I use kube apply task without downloading
kubectl. The error happnens at varidation of
tl.checkPath(this.configfile, 'configfile'); after the kubectl apply
task, I check if these is .kubeconfig or not. It exists.

I guess the problem is asyncronous related problem. That is why, I move
the sesion of the writing .kubeconfig file and make it Sync method.
Now it works.