Adding a PATH and KUBECONFIG for kubectl
To enable istio, we need to add kubectl to PATH and enable to find from
istio. For this puropose, I change the kubectl.vSomeVersion to kubectl.
Also I add PATH and KUBECONFIG environment variables. We can't test it
for this vsts task lib until now. I'll add the test after VSTS-task-lib
subpport the setVariables testing.